Rename logrequests to logconnections
As we now handle inputs and outputs results, rename the package as it
logs the entire connection lifecyle and transit.
